# Conflicts:
#	project/src/Subscriber/Subscriber.java
#	project/src/Subscriber/SubscriberFactory.java
#	project/src/SubscriptionType/SubscriptionType.java
#	project/src/common/SubscriberManager.java
#	project/src/test/SessionTest.java
#	project/src/test/runner.java

+++ project/src/common/Session.java
@@ -115,7 +115,7 @@ public class Session {
 	}
 
 	private int getMinutes(int timeInSeconds) {
-		double dMin = timeInSeconds / 60;
+		double dMin = (double) timeInSeconds / 60;
 		return (int) Math.ceil(dMin);
 	}
